refactor: devided form to components formCardBox and formComponent
