+++ src/components/formCardBox/formCardBox.css
@@ -1,0 +1,9 @@
+.formCardBox {
+  border: 1px solid blueviolet;
+  padding: 10px;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+  width: 100%;
+  min-height: 400px;
+}
