+++ src/components/formComponent/formComponent.css
@@ -1,0 +1,28 @@
+.formsBox {
+  border: 1px solid green;
+  width: 300px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 10px;
+  gap: 3px;
+}
+
+.formsRow {
+  border: 1px solid black;
+  padding: 10px;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.validRow {
+  color: red;
+}
+
+.formsBTN {
+padding: 10px;
+font-size: 25px;
+}
