+++ src/pages/Forms/Forms.css
@@ -6,42 +6,3 @@
   flex-direction: row;
   gap: 10px;
 }
-
-.formsBox {
-  border: 1px solid green;
-  width: 300px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: 10px;
-  gap: 3px;
-}
-
-.formsRow {
-  border: 1px solid black;
-  padding: 10px;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.validRow {
-  color: red;
-}
-
-.formsBTN {
-padding: 10px;
-font-size: 25px;
-}
-
-.formCardBox {
-  border: 1px solid blueviolet;
-  padding: 10px;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px;
-  width: 100%;
-  min-height: 400px;
-}
